opendelta_store/serializers.py
# renders objects into format that can bre rendered into API friendly format

from rest_framework import serializers
from .models import Order


class OrderSerializer(serializers.ModelSerializer):

    class Meta:
        model = Order
        fields = "__all__"
